Fix libcocotbutils on Linux and Mac OS (#2330)
libcocotbutils on Linux and Mac OS depends on the loader library "dl",
however this dependency was never added and cocotb works fine. This is
because libpython loaded libdl.so as a dependency, so it was resolved
transitively.

As of #2268, this dependency on libpython as a load-time dependency was
severed. This continued to not be a porblem because simulators link
against libdl to load VPI library (like cocotb). However, verilator is
different, it is not a AoT compiled simulator that dynamically links in
cocotb, but compiles cocotb support into each simulation, so it doesn't
have the libdl dependency.

The libdl dependency by libcocotbutils is now explicitly added for Linux
and Mac OS. This is not a problem for Windows as all extension modules 
link against kernel32.dll which provides the equivalent dependency on
Windows.